+++ packages/plugin-map/web/src/toolbar/MapSettingsModal.jsx
@@ -9,6 +9,7 @@ import {
   SettingsSection,
   TextInput,
   SettingsPanelFooter,
+  SettingsPanelHeader,
   SearchIcon,
   SettingsMobileHeader,
 } from 'wix-rich-content-ui-components';
@@ -259,15 +260,7 @@ export class MapSettingsModal extends Component {
         )}
 
         <div className={this.styles.map_settings_modal_settings_container} data-hook="mapSettings">
-          <div
-            className={classNames(
-              this.styles.map_settings_modal_title_container,
-              this.styles.map_settings_modal_main_content_block
-            )}
-          >
-            <h3 className={this.styles.map_settings_modal_title}>{t('MapSettings_Title')}</h3>
-          </div>
-
+          <SettingsPanelHeader title={t('MapSettings_Title')} onClose={closeModal} />
           {isMobile
             ? this.renderSettingsSections()
             : wrapWithScrollBars(this.renderSettingsSections())}