+++ packages/tiptap-editor/web/src/components/RicosTiptapEditor/core-configs.ts
@@ -6,6 +6,7 @@ import {
   createHistoryConfig,
   createStylesConfig,
   createUniqueId,
+  createBulletedList,
 } from 'wix-tiptap-extensions';
 
 export const coreConfigs: RicosExtension[] = [
@@ -15,4 +16,5 @@ export const coreConfigs: RicosExtension[] = [
   createFocusConfig(),
   createUniqueId(),
   createCommandsConfig(),
+  createBulletedList(),
 ];